refactor: theme-creation-form descontinuado e a logica esta em cms/theme/create/page assim utilizando metodo DRY
